+++ src/components/Galeria.jsx
@@ -1,69 +1,16 @@
 /*dudas con este componente, podemos hacer la galeria en home*/
-import { NavLink, useNavigate } from "react-router-dom";
-import BookmarkBorderIcon from "@mui/icons-material/BookmarkBorder";
-import IosShareIcon from "@mui/icons-material/IosShare";
-import ChatBubbleOutlineIcon from "@mui/icons-material/ChatBubbleOutline";
+
+import Card from "./Card";
 
 const Galeria = () => {
-	const navigate = useNavigate();
-	const handleClick = (nombre) => {
-		const route = `/home/${nombre}`;
-		navigate(route);
-	};
-
 	return (
 		<>
-			<main className="galeria-container">
-				<div className="galeria">
-					<div className="contenedor-card">
-						<div>
-							<img
-								className="imagen-card"
-								src="../img/pexels-1.jpg"
-								alt="imagenreceta"
-							></img>
-
-							<div className="caja">
-								<h4 className="titulo-caja">NOMBRE RECETA</h4>
-								<NavLink to="/home/:nombre">
-									<button className="btn-caja">
-										Ver más
-									</button>
-								</NavLink>
-							</div>
-						</div>
-						<div className="iconos">
-							<button className="iconos-card">
-								<BookmarkBorderIcon
-									sx={{
-										fontSize: 30,
-										color: "#ffe8d6",
-										margin: "0 1rem",
-									}}
-								/>
-							</button>
-							<button className="iconos-card">
-								<IosShareIcon
-									sx={{
-										fontSize: 30,
-										color: "#ffe8d6",
-										margin: "0 1rem",
-									}}
-								/>
-							</button>
-							<button className="iconos-card">
-								<ChatBubbleOutlineIcon
-									sx={{
-										fontSize: 28,
-										color: "#ffe8d6",
-										margin: "0 1rem",
-									}}
-								/>
-							</button>
-						</div>
-					</div>
-				</div>
-			</main>
+			<div className="galeria">
+				<Card />
+				<Card />
+				<Card />
+				<Card />
+			</div>
 		</>
 	);
 };
